refactor(general): Remove various obsolete code
