+++ lib/core/path.js
@@ -1,7 +1,3 @@
-import d3 from 'd3'
-
-import interpolateLine from '../util/interpolate-line'
-
 /**
  * A path through the network graph. Composed of PathSegments (which
  * are in turn composed of a sequence of graph edges)
@@ -32,27 +28,6 @@ export default class NetworkPath {
     segment.points.forEach(function (point) {
       point.paths.push(this)
     }, this)
-  }
-
-  /** highlight **/
-
-  drawHighlight (display, capExtension) {
-    this.line = d3.svg.line() // the line translation function
-    .x(function (pointInfo, i) {
-      return display.xScale(pointInfo.x) + (pointInfo.offsetX || 0)
-    })
-      .y(function (pointInfo, i) {
-        return display.yScale(pointInfo.y) - (pointInfo.offsetY || 0)
-      })
-      .interpolate(interpolateLine.bind(this))
-
-    this.lineGraph = display.svg.append('path')
-      .attr('id', 'transitive-path-highlight-' + this.parent.getElementId())
-      .attr('class', 'transitive-path-highlight')
-      .style('stroke-width', 24).style('stroke', '#ff4')
-      .style('fill', 'none')
-      .style('visibility', 'hidden')
-      .data([this])
   }
 
   getRenderedSegments () {
