+++ lib/display/canvas-display.js
@@ -1,7 +1,6 @@
-
 import roundedRect from 'rounded-rect'
 
-import Display from './display2'
+import Display from './display'
 
 export default class CanvasDisplay extends Display {
   constructor (transitive) {
