+++ lib/display/display.js
@@ -1,119 +1,94 @@
-const debug = require('debug')('transitive:display')
-
-import d3 from 'd3'
-
-import Legend from './legend'
-import TileLayer from './tile-layer'
-import { sm } from '../util'
-
-/**
- * The D3-based SVG display.
- */
+import linearScale from 'simple-linear-scale'
 
 export default class Display {
-  /**
-   * Display constructor
-   *
-   * @param {Object} options
-   */
-
   constructor (transitive) {
     this.transitive = transitive
-    if (transitive.el) this.el = transitive.el
-    else this.svgContainer = transitive.options.svgContainer
-    this.updateDimensions()
-
-    // Set up the pan/zoom behavior
-    var zoom = this.zoom = d3.behavior.zoom().scaleExtent([0.25, 4])
-
-    const zoomBehavior = () => {
-      this.computeScale()
-      if (this.scale !== this.lastScale) { // zoom action
-        this.zoomChanged()
-      } else { // pan action
-        setTimeout(transitive.refresh.bind(transitive, true), 0)
-      }
-
-      var llb = this.llBounds()
-      debug(`ll bounds: ${llb[0][0]}, ${llb[0][1]} to ${llb[1][0]}, ${llb[1][1]}`)
-    }
-
-    this.zoom.on('zoom.transitive', zoomBehavior)
 
     this.zoomFactors = transitive.options.zoomFactors || this.getDefaultZoomFactors()
 
-    // set up the svg display
+    this.updateActiveZoomFactors(1)
+  }
 
-    if (transitive.options.svg && transitive.options.svgContainer) {
-      this.svg = transitive.options.svg
-    } else {
-      const div = d3.select(this.el).attr('class', 'Transitive')
-      if (transitive.options.zoomEnabled) div.call(zoom)
-      this.svg = div.append('svg')
+  setDimensions (width, height) {
+    this.width = width
+    this.height = height
+  }
+
+  setXDomain (domain) { // [minX , maxX]
+    this.xDomain = domain
+    this.xScale = linearScale(domain, [0, this.width])
+    if (!this.initialXDomain) {
+      this.initialXDomain = domain
+      this.initialXRes = (domain[1] - domain[0]) / this.width
     }
-    this.svg.attr('class', 'schematic-map')
+  }
 
-    // initialize the x/y scale objects
-    this.xScale = d3.scale.linear()
-    this.yScale = d3.scale.linear()
+  setYDomain (domain) { // [minY , maxY]
+    this.yDomain = domain
+    this.yScale = linearScale(domain, [this.height, 0])
+    if (!this.initialYDomain) this.initialYDomain = domain
+  }
 
-    // set up the resize event handler
-    if (transitive.options.autoResize) {
-      d3.select(window).on('resize.display', () => {
-        this.resized()
-        transitive.refresh()
-      })
-    }
+  fitToWorldBounds (bounds) {
+    const domains = this.computeDomainsFromBounds(bounds)
+    this.setXDomain(domains[0])
+    this.setYDomain(domains[1])
+    this.computeScale()
+  }
 
-    // set the scale
-    var bounds
-    if (transitive.options.initialBounds) {
-      bounds = [sm.forward(transitive.options.initialBounds[0]),
-        sm.forward(transitive.options.initialBounds[1])
-      ]
-    } else if (transitive.network && transitive.network.graph) {
-      bounds = transitive.network.graph.bounds()
-    }
-
-    if (bounds) {
-      this.setScale(bounds, transitive.options)
-      this.updateActiveZoomFactors(this.scale)
-      this.lastScale = this.scale
-    } else {
-      this.updateActiveZoomFactors(1)
-    }
-
-    // set up the map layer
-    if (transitive.options.mapboxId) {
-      this.tileLayer = new TileLayer({
-        el: this.el,
-        display: this,
-        graph: transitive.graph,
-        mapboxId: transitive.options.mapboxId
-      })
-    }
-
-    // set up the legend
-    if (transitive.options.legendEl) {
-      this.legend = new Legend(transitive.options.legendEl, this, transitive)
-    }
-
-    transitive.emit('initialize display', transitive, this)
-    return this
+  reset () {
+    this.initialXDomain = null
+    this.initialYDomain = null
+    this.scaleSet = false
+    this.lastScale = undefined
   }
 
   /**
-   * zoomChanged -- called when the zoom level changes, either by through the native
-   * zoom support or the setBounds() API call. Updates zoom factors as needed and
-   * performs appropriate update action (render or refresh)
+   * Apply a transformation {x, y, k} to the *initial* state of the map, where
+   * (x, y) is the pixel offset and k is a scale factor relative to an initial
+   * zoom level of 1.0. Intended primarily to support D3-style panning/zooming.
    */
 
-  zoomChanged () {
-    if (this.updateActiveZoomFactors(this.scale)) {
-      this.transitive.network = null
-      this.transitive.render()
-    } else this.transitive.refresh()
-    this.lastScale = this.scale
+  applyTransform (transform) {
+    const { x, y, k } = transform
+
+    let xMin = this.initialXDomain[0]
+    let xMax = this.initialXDomain[1]
+    let yMin = this.initialYDomain[0]
+    let yMax = this.initialYDomain[1]
+
+    // Apply the scale factor
+    xMax = xMin + (xMax - xMin) / k
+    yMin = yMax - (yMax - yMin) / k
+
+    // Apply the translation
+    const xOffset = -x * (xMax - xMin) / this.width
+    xMin += xOffset
+    xMax += xOffset
+    const yOffset = y * (yMax - yMin) / this.height
+    yMin += yOffset
+    yMax += yOffset
+
+    // Update the scale functions and recompute the internal scale factor
+    this.setXDomain([xMin, xMax])
+    this.setYDomain([yMin, yMax])
+    this.computeScale()
+  }
+
+  getDefaultZoomFactors (data) {
+    return [{
+      minScale: 0,
+      gridCellSize: 25,
+      internalVertexFactor: 1000000,
+      angleConstraint: 45,
+      mergeVertexThreshold: 200
+    }, {
+      minScale: 1.5,
+      gridCellSize: 0,
+      internalVertexFactor: 0,
+      angleConstraint: 5,
+      mergeVertexThreshold: 0
+    }]
   }
 
   updateActiveZoomFactors (scale) {
@@ -134,234 +109,115 @@ export default class Display {
     return updated
   }
 
-  /**
-   * Return default zoom factors
-   */
-
-  getDefaultZoomFactors (data) {
-    return [{
-      minScale: 0,
-      gridCellSize: 25,
-      internalVertexFactor: 1000000,
-      angleConstraint: 45,
-      mergeVertexThreshold: 200
-    }, {
-      minScale: 1.5,
-      gridCellSize: 0,
-      internalVertexFactor: 0,
-      angleConstraint: 5,
-      mergeVertexThreshold: 0
-    }]
-  }
-
-  /**
-   * Empty the display
-   */
-
-  empty () {
-    debug('emptying svg')
-    this.svg.selectAll('*').remove()
-
-    this.haloLayer = this.svg.insert('g', ':first-child')
-  }
-
-  /**
-   * Set the scale
-   */
-
-  setScale (bounds, options) {
-    this.updateDimensions()
-
-    var domains = getDomains(this, this.height, this.width, bounds, options)
-    this.xScale.domain(domains[0])
-    this.yScale.domain(domains[1])
-
-    this.xScale.range([0, this.width])
-    this.yScale.range([this.height, 0])
-
-    debug('x scale %j -> %j', this.xScale.domain(), this.xScale.range())
-    debug('y scale %j -> %j', this.yScale.domain(), this.yScale.range())
-
-    this.zoom
-      .x(this.xScale)
-      .y(this.yScale)
-
-    this.initXRes = (domains[0][1] - domains[0][0]) / this.width
-    this.scale = 1
-
-    this.scaleSet = true
-  }
-
   computeScale () {
-    var newXRes = (this.xScale.domain()[1] - this.xScale.domain()[0]) / this.width
-    this.scale = this.initXRes / newXRes
+    this.lastScale = this.scale
+    this.scaleSet = true
+    const newXRes = (this.xDomain[1] - this.xDomain[0]) / this.width
+    this.scale = this.initialXRes / newXRes
+    if (this.lastScale !== this.scale) this.scaleChanged()
   }
 
-  /** updateDimensions -- compute the width/height of the display */
-
-  updateDimensions () {
-    if (this.el) {
-      this.height = this.el.clientHeight
-      this.width = this.el.clientWidth
-    } else {
-      this.height = this.svgContainer.clientHeight
-      this.width = this.svgContainer.clientWidth
+  scaleChanged () {
+    const zoomFactorsChanged = this.updateActiveZoomFactors(this.scale)
+    if (zoomFactorsChanged) {
+      this.transitive.network = null
+      this.transitive.render()
     }
   }
 
   /**
-   * updateDomains -- set x/y domains of geographic (spherical mercator) coordinate
-   * system. Does *not* check/adjust aspect ratio.
+   * Compute the x/y coordinate space domains to fit the graph.
    */
 
-  updateDomains (bounds) {
-    this.xScale.domain([bounds[0][0], bounds[1][0]])
-    this.yScale.domain([bounds[0][1], bounds[1][1]])
+  computeDomainsFromBounds (bounds) {
+    var xmin = bounds[0][0]
+    var xmax = bounds[1][0]
+    var ymin = bounds[0][1]
+    var ymax = bounds[1][1]
+    var xRange = xmax - xmin
+    var yRange = ymax - ymin
 
-    this.zoom
-      .x(this.xScale)
-      .y(this.yScale)
+    const { options } = this.transitive
 
-    this.computeScale()
-  }
+    var paddingFactor = (options && options.paddingFactor)
+      ? options.paddingFactor
+      : 0.1
 
-  resized () {
-    var newWidth = this.el ? this.el.clientWidth : this.svgContainer.clientWidth
-    var newHeight = this.el ? this.el.clientHeight : this.svgContainer.clientHeight
+    var margins = this.getMargins()
 
-    var xDomain = this.xScale.domain()
-    var xFactor = newWidth / this.width
-    var xDomainAdj = (xDomain[1] - xDomain[0]) * (xFactor - 1) / 2
-    this.xScale.domain([xDomain[0] - xDomainAdj, xDomain[1] + xDomainAdj])
+    var usableHeight = this.height - margins.top - margins.bottom
+    var usableWidth = this.width - margins.left - margins.right
+    var displayAspect = this.width / this.height
+    var usableDisplayAspect = usableWidth / usableHeight
+    var graphAspect = xRange / (yRange === 0 ? -Infinity : yRange)
 
-    var yDomain = this.yScale.domain()
-    var yFactor = newHeight / this.height
-    var yDomainAdj = (yDomain[1] - yDomain[0]) * (yFactor - 1) / 2
-    this.yScale.domain([yDomain[0] - yDomainAdj, yDomain[1] + yDomainAdj])
+    var padding
+    var dispX1, dispX2, dispY1, dispY2
+    var dispXRange, dispYRange
 
-    this.xScale.range([0, newWidth])
-    this.yScale.range([newHeight, 0])
+    if (usableDisplayAspect > graphAspect) { // y-axis is limiting
+      padding = paddingFactor * yRange
+      dispY1 = ymin - padding
+      dispY2 = ymax + padding
+      dispYRange = yRange + 2 * padding
+      var addedYRange = (this.height / usableHeight * dispYRange) - dispYRange
+      if (margins.top > 0 || margins.bottom > 0) {
+        dispY1 -= margins.bottom / (margins.bottom + margins.top) * addedYRange
+        dispY2 += margins.top / (margins.bottom + margins.top) * addedYRange
+      }
+      dispXRange = (dispY2 - dispY1) * displayAspect
+      var xOffset = (margins.left - margins.right) / this.width
+      var xMidpoint = (xmax + xmin - dispXRange * xOffset) / 2
+      dispX1 = xMidpoint - dispXRange / 2
+      dispX2 = xMidpoint + dispXRange / 2
+    } else { // x-axis limiting
+      padding = paddingFactor * xRange
+      dispX1 = xmin - padding
+      dispX2 = xmax + padding
+      dispXRange = xRange + 2 * padding
+      var addedXRange = (this.width / usableWidth * dispXRange) - dispXRange
+      if (margins.left > 0 || margins.right > 0) {
+        dispX1 -= margins.left / (margins.left + margins.right) * addedXRange
+        dispX2 += margins.right / (margins.left + margins.right) * addedXRange
+      }
 
-    this.height = newHeight
-    this.width = newWidth
+      dispYRange = (dispX2 - dispX1) / displayAspect
+      var yOffset = (margins.bottom - margins.top) / this.height
+      var yMidpoint = (ymax + ymin - dispYRange * yOffset) / 2
+      dispY1 = yMidpoint - dispYRange / 2
+      dispY2 = yMidpoint + dispYRange / 2
+    }
 
-    this.zoom
-      .x(this.xScale)
-      .y(this.yScale)
-  }
-
-  xyBounds () {
-    var x = this.xScale.domain()
-    var y = this.yScale.domain()
     return [
-      [x[0], y[0]],
-      [x[1], y[1]]
+      [dispX1, dispX2],
+      [dispY1, dispY2]
     ]
   }
 
-  /**
-   * Lat/lon bounds
-   */
-
-  llBounds () {
-    var x = this.xScale.domain()
-    var y = this.yScale.domain()
-
-    return [
-      sm.inverse([x[0], y[0]]),
-      sm.inverse([x[1], y[1]])
-    ]
+  getMargins () {
+    return Object.assign({
+      left: 0,
+      right: 0,
+      top: 0,
+      bottom: 0
+    }, this.transitive.options.displayMargins)
   }
 
   isInRange (x, y) {
-    var xRange = this.xScale.range()
-    var yRange = this.yScale.range()
-
-    return x >= xRange[0] && x <= xRange[1] && y >= yRange[1] && y <= yRange[0]
-  }
-}
-
-/**
- * Compute the x/y coordinate space domains to fit the graph.
- */
-
-function getDomains (display, height, width, bounds, options) {
-  var xmin = bounds[0][0]
-  var xmax = bounds[1][0]
-  var ymin = bounds[0][1]
-  var ymax = bounds[1][1]
-  var xRange = xmax - xmin
-  var yRange = ymax - ymin
-
-  var paddingFactor = (options && options.paddingFactor)
-    ? options.paddingFactor
-    : 0.1
-
-  var margins = getMargins(options)
-
-  var usableHeight = height - margins.top - margins.bottom
-  var usableWidth = width - margins.left - margins.right
-  var displayAspect = width / height
-  var usableDisplayAspect = usableWidth / usableHeight
-  var graphAspect = xRange / (yRange === 0 ? -Infinity : yRange)
-
-  var padding
-  var dispX1, dispX2, dispY1, dispY2
-  var dispXRange, dispYRange
-
-  if (usableDisplayAspect > graphAspect) { // y-axis is limiting
-    padding = paddingFactor * yRange
-    dispY1 = ymin - padding
-    dispY2 = ymax + padding
-    dispYRange = yRange + 2 * padding
-    var addedYRange = (height / usableHeight * dispYRange) - dispYRange
-    if (margins.top > 0 || margins.bottom > 0) {
-      dispY1 -= margins.bottom / (margins.bottom + margins.top) * addedYRange
-      dispY2 += margins.top / (margins.bottom + margins.top) * addedYRange
-    }
-    dispXRange = (dispY2 - dispY1) * displayAspect
-    var xOffset = (margins.left - margins.right) / width
-    var xMidpoint = (xmax + xmin - dispXRange * xOffset) / 2
-    dispX1 = xMidpoint - dispXRange / 2
-    dispX2 = xMidpoint + dispXRange / 2
-  } else { // x-axis limiting
-    padding = paddingFactor * xRange
-    dispX1 = xmin - padding
-    dispX2 = xmax + padding
-    dispXRange = xRange + 2 * padding
-    var addedXRange = (width / usableWidth * dispXRange) - dispXRange
-    if (margins.left > 0 || margins.right > 0) {
-      dispX1 -= margins.left / (margins.left + margins.right) * addedXRange
-      dispX2 += margins.right / (margins.left + margins.right) * addedXRange
-    }
-
-    dispYRange = (dispX2 - dispX1) / displayAspect
-    var yOffset = (margins.bottom - margins.top) / height
-    var yMidpoint = (ymax + ymin - dispYRange * yOffset) / 2
-    dispY1 = yMidpoint - dispYRange / 2
-    dispY2 = yMidpoint + dispYRange / 2
+    return x >= 0 && x <= this.width && y >= 0 && y <= this.height
   }
 
-  return [
-    [dispX1, dispX2],
-    [dispY1, dispY2]
-  ]
-}
+  /** Methods to be defined by subclasses **/
 
-function getMargins (options) {
-  var margins = {
-    left: 0,
-    right: 0,
-    top: 0,
-    bottom: 0
-  }
+  clear () { }
 
-  if (options && options.displayMargins) {
-    if (options.displayMargins.top) margins.top = options.displayMargins.top
-    if (options.displayMargins.bottom) margins.bottom = options.displayMargins.bottom
-    if (options.displayMargins.left) margins.left = options.displayMargins.left
-    if (options.displayMargins.right) margins.right = options.displayMargins.right
-  }
+  drawCircle (coord, attrs) { }
 
-  return margins
+  drawEllipse (coord, attrs) { }
+
+  drawRect (upperLeft, attrs) { }
+
+  drawText (text, anchor, attrs) { }
+
+  drawPath (pathStr, attrs) { }
 }
