+++ lib/display/svg-display.js
@@ -1,4 +1,4 @@
-import Display from './display2'
+import Display from './display'
 import SVG from 'svg.js'
 
 export default class SvgDisplay extends Display {
@@ -10,8 +10,6 @@ export default class SvgDisplay extends Display {
 
     // We have a DOM element
     if (el) {
-      console.log('--> setting el in SvgDisplay');
-
       this.setDimensions(el.clientWidth, el.clientHeight)
       this.svg = SVG(el)
     }
