+++ lib/point/multipoint.js
@@ -152,17 +152,7 @@ export default class MultiPoint extends Point {
 
     if (!this.renderData) return
 
-    // set up the main svg group for this stop
-    /*this.initSvg(display)
-    this.svgGroup
-      .attr('class', 'transitive-sortable')
-      .datum({
-        owner: this,
-        sortableType: 'POINT_MULTI'
-      })*/
-
-    //if (this.containsSegmentEndPoint()) this.initMergedMarker(display)
-
+    // Compute the bounds of the merged marker
     const xArr = this.renderData.map(d => d.x)
     const yArr = this.renderData.map(d => d.y)
     const xMin = Math.min(...xArr)
@@ -176,6 +166,7 @@ export default class MultiPoint extends Point {
     const width = xMax - xMin + r * 2
     const height = yMax - yMin + r * 2
 
+    // Draw the merged marker
     display.drawRect({ x, y }, {
       width,
       height,
@@ -186,15 +177,10 @@ export default class MultiPoint extends Point {
       'stroke-width': 2
     })
 
-    // store marker bounding box
+    // Store marker bounding box
     this.markerBBox = { x, y, width, height }
 
-    // set up the pattern markers
-    /* this.marker = this.markerSvg.selectAll('circle')
-      .data(this.renderData)
-      .enter()
-      .append('circle')
-      .attr('class', 'transitive-multipoint-marker-pattern'); */
+    // TODO: support pattern-specific markers
   }
 
   initMergedMarker (display) {
@@ -212,39 +198,6 @@ export default class MultiPoint extends Point {
         })
         .attr('class', 'transitive-multipoint-marker-merged')
     }
-  }
-
-  /**
-   * Refresh the point
-   *
-   * @param {Display} display
-   */
-
-  refresh (display) {
-    /*if (!this.renderData) return
-
-    // refresh the merged marker
-    if (this.mergedMarker) {
-      if (!this.mergedMarkerData) this.initMarkerData(display)
-
-      this.mergedMarker.datum({
-        owner: this
-      })
-      for (const key in this.mergedMarkerData) {
-        this.mergedMarker.attr(key, this.mergedMarkerData[key])
-      }
-    }*/
-
-    /* var cx, cy;
-    // refresh the pattern-level markers
-    this.marker.data(this.renderData);
-    this.marker.attr('transform', function (d, i) {
-      cx = d.x;
-      cy = d.y;
-      var x = display.xScale(d.x) + d.offsetX;
-      var y = display.yScale(d.y) - d.offsetY;
-      return 'translate(' + x +', ' + y +')';
-    }); */
   }
 
   getRenderDataArray () {
