+++ lib/point/place.js
@@ -1,5 +1,3 @@
-import d3 from 'd3'
-
 import Point from './point'
 import { sm } from '../util'
 
@@ -128,32 +126,6 @@ export default class Place extends Point {
       width: radius * 2,
       height: radius * 2
     }
-    /*this.initSvg(display)
-    this.svgGroup
-      .attr('class', 'transitive-sortable')
-      .datum({
-        owner: this,
-        sortableType: 'POINT_PLACE'
-      })
-
-    // set up the markers
-    this.marker = this.markerSvg.append('circle')
-      .datum({
-        owner: this
-      })
-      .attr('class', 'transitive-place-circle')
-
-    var iconUrl = display.styler.compute(display.styler.places_icon['xlink:href'], display, {
-      owner: this
-    })
-    if (iconUrl) {
-      this.icon = this.markerSvg.append('image')
-        .datum({
-          owner: this
-        })
-        .attr('class', 'transitive-place-icon')
-        .attr('xlink:href', iconUrl)
-    }*/
   }
 
   /**
@@ -163,54 +135,5 @@ export default class Place extends Point {
    */
 
   refresh (display) {
-    /*if (!this.renderData) return
-
-    // compute the x/y location -- use the real-world coordinates if
-    // geographic rendering is enabled, otherwise use snapped vertex coords
-    var x = display.xScale(
-      display.activeZoomFactors.useGeographicRendering
-        ? this.worldX
-        : this.graphVertex.x
-    )
-    var y = display.yScale(
-      display.activeZoomFactors.useGeographicRendering
-        ? this.worldY
-        : this.graphVertex.y
-    )
-    const translate = 'translate(' + x + ', ' + y + ')'
-
-    // update the marker/icon translations
-    this.marker.attr('transform', translate)
-    if (this.icon) this.icon.attr('transform', translate)*/
-  }
-
-  makeDraggable (transitive) {
-    /*var place = this
-    var display = transitive.display
-    var drag = d3.behavior.drag()
-      .on('dragstart', function () {
-        d3.event.sourceEvent.stopPropagation() // silence other listeners
-      })
-      .on('drag', function () {
-        if (place.graphVertex) {
-          var boundingRect = display.el.getBoundingClientRect()
-          var x = display.xScale.invert(d3.event.sourceEvent.pageX -
-            boundingRect.left)
-          var y = display.yScale.invert(d3.event.sourceEvent.pageY -
-            boundingRect.top)
-
-          place.worldX = x
-          place.worldY = y
-          var ll = sm.inverse([x, y])
-          place.place_lon = ll[0]
-          place.place_lat = ll[1]
-
-          place.refresh(display)
-        }
-      })
-      .on('dragend', function () {
-        transitive.emit('place.' + place.getId() + '.dragend', place)
-      })
-    this.markerSvg.call(drag)*/
   }
 }
