+++ lib/point/point.js
@@ -91,17 +91,7 @@ export default class Point {
    * @param {Display} display
    */
 
-  render (display) {
-    //this.label.svgGroup = null
-  }
-
-  /**
-   * Refresh a previously drawn point
-   *
-   * @param {Display} display
-   */
-
-  refresh (display) {}
+  render (display) {}
 
   addRenderData () {}
 
@@ -114,17 +104,6 @@ export default class Point {
   containsToPoint () {
     return false
   }
-
-  /*initSvg (display) {
-    // set up the main svg group for this stop
-    this.svgGroup = display.svg.append('g')
-      .attr('id', `transitive-${this.getType().toLowerCase()}-${this.getId()}`)
-        // .attr('class', 'transitive-sortable')
-        .datum(this)
-
-    this.markerSvg = this.svgGroup.append('g')
-    this.labelSvg = this.svgGroup.append('g')
-  }*/
 
   //* * Shared geom utility functions **//
 
@@ -236,11 +215,6 @@ export default class Point {
       }
     }
   }
-
-  /*drawLabel (display) {
-    if (!this.renderLabel) return
-    this.label.draw(display)
-  }*/
 
   getMarkerBBox () {
     return this.markerBBox
