+++ lib/point/stop.js
@@ -152,30 +152,11 @@ export default class Stop extends Point {
 
     if (this.patternCount === 0) return
 
-    //if (!this.mergedMarkerData) this.initMarkerData(display)
-
     this.initMarkerData(display)
-    // refresh the pattern-level markers
-    /*this.patternMarkers.data(this.getRenderDataArray())
-
-    this.patternMarkers.attr('transform', (d, i) => {
-      if (!isNaN(d.x) && !isNaN(d.y)) {
-        var x = d.x + this.placeOffsets.x
-        var y = d.y + this.placeOffsets.y
-        return `translate(${x}, ${y})`
-      }
-    })
-
-    // refresh the merged marker
-    if (this.mergedMarker) {
-      this.mergedMarker.datum(this.getMergedRenderData())
-      if (!isNaN(this.mergedMarkerData.x) && !isNaN(this.mergedMarkerData.y)) {
-        for (const key in this.mergedMarkerData) this.mergedMarker.attr(key, this.mergedMarkerData[key])
-      }
-    }*/
 
     const styler = display.styler
 
+    // For segment endpoints, draw the "merged" marker
     if (this.isSegmentEndPoint && this.mergedMarkerData) {
       display.drawRect({
         x: this.mergedMarkerData.x,
@@ -212,63 +193,7 @@ export default class Stop extends Point {
         })
       }*/
     }
-
-    /*if (Object.keys(this.patternRenderData).length === 0) return
-
-    var renderDataArray = this.getRenderDataArray()
-
-    this.initSvg(display)
-
-    // set up the merged marker
-    this.mergedMarker = this.markerSvg.append('g').append('rect')
-      .attr('class', 'transitive-sortable transitive-stop-marker-merged')
-      .datum(this.getMergedRenderData())
-
-    // set up the pattern-specific markers
-    this.patternMarkers = this.markerSvg.append('g').selectAll('circle')
-      .data(renderDataArray)
-      .enter()
-      .append('circle')
-      .attr('class', 'transitive-sortable transitive-stop-marker-pattern')*/
   }
-
-  /**
-   * Refresh the stop
-   *
-   * @param {Display} display
-   */
-
-  refresh (display) {
-    /*if (this.patternCount === 0) return
-
-    if (!this.mergedMarkerData) this.initMarkerData(display)
-
-    // refresh the pattern-level markers
-    this.patternMarkers.data(this.getRenderDataArray())
-
-    this.patternMarkers.attr('transform', (d, i) => {
-      if (!isNaN(d.x) && !isNaN(d.y)) {
-        var x = d.x + this.placeOffsets.x
-        var y = d.y + this.placeOffsets.y
-        return `translate(${x}, ${y})`
-      }
-    })
-
-    // refresh the merged marker
-    if (this.mergedMarker) {
-      this.mergedMarker.datum(this.getMergedRenderData())
-      if (!isNaN(this.mergedMarkerData.x) && !isNaN(this.mergedMarkerData.y)) {
-        for (const key in this.mergedMarkerData) this.mergedMarker.attr(key, this.mergedMarkerData[key])
-      }
-    }*/
-  }
-
-  /*getMergedRenderData () {
-    return {
-      owner: this,
-      sortableType: 'POINT_STOP_MERGED'
-    }
-  }*/
 
   getRenderDataArray () {
     var dataArray = []
@@ -276,30 +201,6 @@ export default class Stop extends Point {
       dataArray.push(this.patternRenderData[patternId])
     }
     return dataArray
-  }
-
-  /*isFocused () {
-    if (this.mergedMarker || !this.patternRenderData) {
-      return (this.focused === true)
-    }
-
-    var focused = true
-    for (var patternId in this.patternRenderData) {
-      focused = this && this.isPatternFocused(patternId)
-    }
-    return focused
-  }*/
-
-  runFocusTransition (display, callback) {
-    if (this.mergedMarker) {
-      var newStrokeColor = display.styler.compute(display.styler.stops_merged
-        .stroke, display, {
-          owner: this
-        })
-      this.mergedMarker.transition().style('stroke', newStrokeColor).call(
-        callback)
-    }
-    if (this.label) this.label.runFocusTransition(display, callback)
   }
 
   clearRenderData () {
