+++ lib/renderer/default-renderer.js
@@ -84,70 +84,6 @@ export default class DefaultRenderer extends Renderer {
     // if (display.legend) display.legend.render(legendSegments)
   }
 
-  /**
-   * Refresh
-   */
-
-  refresh (panning) {
-    /*super.refresh(panning)
-
-    var display = this.transitive.display
-    var network = this.transitive.network
-    var styler = this.transitive.styler
-
-    forEach(network.graph.vertices, vertex => {
-      vertex.point.clearRenderData()
-    })
-    forEach(network.graph.edges, edge => {
-      edge.clearRenderData()
-    })
-
-    // refresh the segment and point marker data
-    this.refreshSegmentRenderData()
-    forEach(network.graph.vertices, vertex => {
-      vertex.point.initMarkerData(display)
-    })
-
-    this.renderedSegments = []
-    forEach(network.paths, path => {
-      forEach(path.segments, pathSegment => {
-        forEach(pathSegment.renderedSegments, rSegment => {
-          rSegment.refresh(display)
-          this.renderedSegments.push(rSegment)
-        })
-      })
-    })
-
-    forEach(network.graph.vertices, vertex => {
-      var point = vertex.point
-      if (!point.svgGroup) return // check if this point is not currently rendered
-      styler.stylePoint(display, point)
-      point.refresh(display)
-    })
-
-    // re-draw the edge-based points
-    forEach(network.graph.edges, edge => {
-      forEach(edge.pointArray, point => {
-        if (!point.svgGroup) return // check if this point is not currently rendered
-        styler.styleStop(display, point)
-        point.refresh(display)
-      })
-    })
-
-    // refresh the label layout
-    var labeledElements = this.transitive.labeler.doLayout()
-    forEach(labeledElements.points, point => {
-      point.refreshLabel(display)
-      styler.stylePointLabel(display, point)
-    })
-    forEach(this.transitive.labeler.segmentLabels, label => {
-      label.refresh(display)
-      styler.styleSegmentLabel(display, label)
-    })
-
-    this.sortElements()*/
-  }
-
   refreshSegmentRenderData () {
     forEach(this.transitive.network.renderedEdges, rEdge => {
       rEdge.refreshRenderData(this.transitive.display)
