+++ lib/renderer/renderedsegment.js
@@ -1,6 +1,5 @@
 import { forEach } from 'lodash'
 
-import interpolateLine from '../util/interpolate-line'
 import { distance, renderDataToSvgPath } from '../util'
 
 /**
@@ -58,36 +57,6 @@ export default class RenderedSegment {
       'stroke-dasharray': styler.compute2('segments', 'stroke-dasharray', this),
       'stroke-linecap': styler.compute2('segments', 'stroke-linecap', this)
     })
-  }
-
-  refresh (display, renderData) {
-    /*if (renderData) {
-      this.renderData = renderData
-    } else {
-      this.renderData = []
-      forEach(this.renderedEdges, rEdge => {
-        this.renderData = this.renderData.concat(rEdge.renderData)
-      })
-    }
-
-    // Check if this segment is to be drawn as an arc; if so replace render data
-    if (this.pathSegment.journeySegment && this.pathSegment.journeySegment.arc) {
-      const first = this.renderData[0]
-      const last = this.renderData[this.renderData.length - 1]
-      const arc = {
-        x: last.x,
-        y: last.y,
-        arc: -45,
-        radius: distance(first.x, first.y, last.x, last.y)
-      }
-      this.renderData = [first, arc, last]
-    }
-
-    var lineData = this.line(this.renderData)
-    this.lineGraph.attr('d', lineData)
-    this.lineGraphFront.attr('d', lineData)
-    if (this.lineGraphHalo) this.lineGraphHalo.attr('d', lineData)
-    display.styler.styleSegment(display, this)*/
   }
 
   setFocused (focused) {
