+++ lib/renderer/renderer.js
@@ -17,31 +17,22 @@ export default class Renderer {
   }
 
   render () {
-    var display = this.transitive.display
+    const display = this.transitive.display
+    const graph = this.transitive.network.graph
+
     display.styler = this.transitive.styler
 
-    // remove all old svg elements
-    display.clear()
-  }
-
-  /**
-   * Refresh
-   */
-
-  refresh (panning) {
-    var display = this.transitive.display
-    var network = this.transitive.network
-
-    if (display.tileLayer) display.tileLayer.zoomed()
-
-    network.graph.vertices.forEach(function (vertex) {
+    graph.vertices.forEach(vertex => {
       vertex.point.clearRenderData()
     })
-    network.graph.edges.forEach(function (edge) {
+    graph.edges.forEach(edge => {
       edge.clearRenderData()
     })
 
-    // draw the grid, if necessary
+    // Clear the display
+    display.clear()
+
+    // Draw the grid, if necessary
     if (this.transitive.options.drawGrid) drawGrid(display, this.gridCellSize)
   }
