+++ lib/renderer/wireframe-renderer.js
@@ -41,6 +41,4 @@ export default class WireframeRenderer extends Renderer {
       })
     })
   }
-
-  refresh (panning) { }
 }
