+++ lib/styler/styler.js
@@ -108,230 +108,6 @@ export default class Styler {
     }
   }
 
-  /**
-   * Style a Segment using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {RenderedSegment} Transitive RenderedSegment object
-   */
-
-  styleSegment (display, segment) {
-    if (segment.lineGraphHalo) {
-      this.applyAttrAndStyle(
-        display,
-        segment.lineGraphHalo,
-        this.segments_halo
-      )
-    }
-
-    this.applyAttrAndStyle(
-      display,
-      segment.lineGraph,
-      this.segments
-    )
-
-    this.applyAttrAndStyle(
-      display,
-      segment.lineGraphFront,
-      this.segments_front
-    )
-  }
-
-  /**
-   * Style a WireframeEdge using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {WireframeEdge} Transitive WireframeEdge object
-   */
-
-  styleWireframeEdge (display, wfEdge) {
-    this.applyAttrAndStyle(
-      display,
-      wfEdge.svgGroup.selectAll('.transitive-wireframe-edge-line'),
-      this.wireframe_edges
-    )
-  }
-
-  /**
-   * Style a Point using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {Point} Transitive Point object
-   */
-
-  stylePoint (display, point) {
-    if (point.getType() === 'STOP') this.styleStop(display, point)
-    if (point.getType() === 'PLACE') this.stylePlace(display, point)
-    if (point.getType() === 'MULTI') this.styleMultiPoint(display, point)
-    if (point.getType() === 'WIREFRAME_VERTEX') this.styleWireframeVertex(display, point)
-  }
-
-  /**
-   * Style a Stop using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {Stop} Transitive Stop object
-   */
-
-  styleStop (display, stop) {
-    this.applyAttrAndStyle(
-      display,
-      stop.patternMarkers,
-      this.stops_pattern
-    )
-
-    this.applyAttrAndStyle(
-      display,
-      stop.mergedMarker,
-      this.stops_merged
-    )
-
-    this.applyAttrAndStyle(
-      display,
-      stop.svgGroup.selectAll('.transitive-stop-label'),
-      this.labels
-    )
-  }
-
-  /**
-   * Style a Place using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {Place} Transitive Place object
-   */
-
-  stylePlace (display, place) {
-    this.applyAttrAndStyle(
-      display,
-      place.svgGroup.selectAll('.transitive-place-circle'),
-      this.places
-    )
-
-    this.applyAttrAndStyle(
-      display,
-      place.svgGroup.selectAll('.transitive-place-icon'),
-      this.places_icon
-    )
-
-    this.applyAttrAndStyle(
-      display,
-      place.svgGroup.selectAll('.transitive-place-label'),
-      this.labels
-    )
-  }
-
-  /**
-   * Style a MultiPoint using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {MultiPoint} Transitive MultiPoint object
-   */
-
-  styleMultiPoint (display, multipoint) {
-    this.applyAttrAndStyle(
-      display,
-      multipoint.svgGroup.selectAll('.transitive-multipoint-marker-pattern'),
-      this.multipoints_pattern
-    )
-
-    this.applyAttrAndStyle(
-      display,
-      multipoint.svgGroup.selectAll('.transitive-multipoint-marker-merged'),
-      this.multipoints_merged
-    )
-
-    this.applyAttrAndStyle(
-      display,
-      multipoint.svgGroup.selectAll('.transitive-multi-label'),
-      this.labels
-    )
-  }
-
-  /**
-   * Style a WireframeVertex using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {WireframeVertex} Transitive WireframeVertex object
-   */
-
-  styleWireframeVertex (display, wfVertex) {
-    this.applyAttrAndStyle(
-      display,
-      wfVertex.svgGroup.selectAll('.transitive-wireframe-vertex-circle'),
-      this.wireframe_vertices
-    )
-  }
-
-  /**
-   * Style a Point label using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {Point} Transitive Point object
-   */
-
-  stylePointLabel (display, point) {
-    var pointType = point.getType().toLowerCase()
-
-    this.applyAttrAndStyle(
-      display,
-      point.svgGroup.selectAll('.transitive-' + pointType + '-label'),
-      this.labels
-    )
-  }
-
-  /**
-   * Style a Segment label using the rules defined in styles.js or the Transitive options
-   *
-   * @param {Display} Transitive Display object
-   * @param {SegmentLabel} Transitive SegmentLabel object
-   */
-
-  styleSegmentLabel (display, label) {
-    this.applyAttrAndStyle(
-      display,
-      label.svgGroup.selectAll('.transitive-segment-label-container'),
-      this.segment_label_containers
-    )
-    this.applyAttrAndStyle(
-      display,
-      label.svgGroup.selectAll('.transitive-segment-label'),
-      this.segment_labels
-    )
-  }
-
-  /**
-   * Check if it's an attribute or a style and apply accordingly
-   *
-   * @param {Display} the Display object
-   * @param {Object} a D3 list of elements
-   * @param {Object} the list of attributes
-   */
-
-  applyAttrAndStyle (display, elements, attributes) {
-    for (var name in attributes) {
-      var rules = attributes[name]
-      var fn = svgAttributes.indexOf(name) === -1 ? 'style' : 'attr'
-
-      this.applyRules(display, elements, name, rules, fn)
-    }
-  }
-
-  /**
-   * Apply style/attribute rules to a list of elements
-   *
-   * @param {Display} display object
-   * @param {Object} elements
-   * @param {String} rule name
-   * @param {Array} rules
-   * @param {String} style/attr
-   */
-
-  applyRules (display, elements, name, rules, fn) {
-    var self = this
-    elements[fn](name, function (data, index) {
-      return self.compute(rules, display, data, index)
-    })
-  }
 
   /**
    * Compute a style rule based on the current display and data
