+++ lib/transitive.js
@@ -2,11 +2,9 @@ import d3 from 'd3'
 import Emitter from 'component-emitter'
 
 import Network from './core/network'
-//import Display from './display/display'
 import SvgDisplay from './display/svg-display'
 import CanvasDisplay from './display/canvas-display'
 import DefaultRenderer from './renderer/default-renderer'
-//import CanvasRenderer from './renderer/canvas-renderer'
 import WireframeRenderer from './renderer/wireframe-renderer'
 import Styler from './styler/styler'
 import Labeler from './labeler/labeler'
@@ -104,24 +102,6 @@ export default class Transitive {
 
     this.el = el
 
-    /*this.display = this.options.display === 'canvas'
-      ? new CanvasDisplay(this)
-      : new SvgDisplay(this)*/
-
-    // Emit click events
-    /*var self = this
-    this.display.svg.on('click', function () {
-      var x = d3.event.x
-      var y = d3.event.y
-      var geographic = sm.inverse([x, y])
-      self.emit('click', {
-        x: x,
-        y: y,
-        lng: geographic[0],
-        lat: geographic[1]
-      })
-    })*/
-
     this.emit('set element', this, this.el)
     return this
   }
@@ -171,18 +151,6 @@ export default class Transitive {
 
     this.emit('render to', this)
     return this
-  }
-
-  /**
-   * Refresh
-   */
-
-  refresh (panning) {
-    if (!this.network) {
-      this.render()
-    }
-
-    this.renderer.refresh()
   }
 
   /**
